+++ components/list-group/examples/00-default.js
@@ -67,6 +67,7 @@ function Example({ brand }) {
 					'List item 2',
 					'List item 3',
 					'List item 4',
+					// eslint-disable-next-line react/jsx-key
 					<Fragment>
 						Send me sms reminders
 						<ButtonGroup size="small" name="example-default-1" css={styleExample}>
@@ -74,6 +75,7 @@ function Example({ brand }) {
 							<BtnGroupItem value="no">No</BtnGroupItem>
 						</ButtonGroup>
 					</Fragment>,
+					// eslint-disable-next-line react/jsx-key
 					<Fragment>
 						Call me all the time
 						<ButtonGroup size="small" name="example-default-3" css={styleExample}>