+++ website/src/pages/_app.js
@@ -34,11 +34,12 @@ const GELApp = ({ Component, pageProps, apollo, brand }) => {
 		return () => {
 			window.removeEventListener('scroll', setScrollClass);
 		};
+		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, []);
 
 	return (
 		<Fragment>
-			{/* <Script
+			<Script
 				src="https://www.googletagmanager.com/gtag/js?id=GTM-TKBQFPG"
 				strategy="afterInteractive"
 			/>
@@ -50,7 +51,7 @@ const GELApp = ({ Component, pageProps, apollo, brand }) => {
 
           gtag('config', 'GTM-TKBQFPG');
         `}
-			</Script> */}
+			</Script>
 			<ApolloProvider client={apolloClient}>
 				<Layout {...pageProps} brand={brand}>
 					<Component {...pageProps} />