+++ components/panel/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @westpac/panel
 
+## 2.1.0
+
+### Minor Changes
+
+- 39e90efa: Updated components to export TypeScript interfaces using existing prop types.
+
+### Patch Changes
+
+- Updated dependencies [39e90efa]
+- Updated dependencies [8d3f5b3d]
+  - @westpac/body@2.1.0
+  - @westpac/core@3.1.0
+  - @westpac/heading@2.1.0
+
 ## 2.0.0
 
 ### Major Changes